+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# 0.4.5
+
+## Changed
+
+* Replace `winapi` dependency with `windows-sys`
+  (https://github.com/rust-lang/socket2/pull/303).
+
+## Added
+
+* `Socket::join_ssm_v4` and `Socket::leave_ssm_v4`
+  (https://github.com/rust-lang/socket2/pull/298).
+* `Socket::set_recv_tos` and `Socket::recv_tos`
+  (https://github.com/rust-lang/socket2/pull/299).
+
+## Fixed
+
+* Fix OpenBSD build
+  (https://github.com/rust-lang/socket2/pull/291).
+* Fix OpenBSD build
+  (https://github.com/rust-lang/socket2/pull/291).
+
 # 0.4.4
 
 ## Fixed
@@ -169,7 +190,6 @@
 * Use correct inmutable references in `Socket::send_with_flags` and
   `Socket::send_out_of_band`.
 * Use `IPPROTO_IPV6` in `Socket::join_multicast_v6` on Windows.
-* Don't assume the memory layout of `std::net::SocketAddr`.
 * Use `c_int` instead of `i32` where appropriate.
 
 ## From v0.4.0-alpha.1 to v0.4.0-alpha.2
@@ -190,3 +210,8 @@
 ## From v0.4.0-alpha.5 to v0.4.0
 
 * Use `SO_LINGER_SEC` on macOS for `Socket::get/set_linger`.
+
+# 0.3.16
+
+* Don't assume the memory layout of `std::net::SocketAddr`.
+* Other changes omited